requirements.txt
streamlit
pandas
requests
altair
python-dateutil
